tdd-katas/report-generator-kata/report-generator-python/pyproject.toml
[tool.poetry]
name = "ReportGeneration"
version = "0.1.0"
description = "Report generation kata"
authors = ["Zeger Hendrikse  <zegerh@yahoo.co.uk>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
pytest-watch = "^4.2.0"
pytest = "^7.3.2"
pyhamcrest = "^2.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
